% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bas.R
\name{rmp_create_bas}
\alias{rmp_create_bas}
\title{Create a RMODPATH mpbas object}
\usage{
rmp_create_bas(
  dis,
  hnoflo,
  hdry,
  packagelabel = NULL,
  defaultifacevalue = NULL,
  laytyp,
  ibound,
  porosity,
  bas = NULL,
  flow = NULL,
  modflow = NULL
)
}
\arguments{
\item{dis}{\code{RMODFLOW} dis object, required to obtain dimensions of ibound and porosity arrays.}

\item{hnoflo}{numeric; value of head used in the MODFLOW simulation to represent inactive cells}

\item{hdry}{numeric; values of thead used in the MODFLOW simulation to represent dry cells}

\item{packagelabel}{optional character vector; text string used by MODFLOW in the budget output file to label flow rates for a stress package. Should have the same length as defaultifacevalue.}

\item{defaultifacevalue}{optional integer vector; with the IFACE value of the corresponding packagelabel stress package. Value must be in range of 0 to 6. Should have the same length as packagelabel.}

\item{laytyp}{numeric vector with length equal to the number of MODFLOW layers; specifies the layertype for each MODFLOW layer. 0 = confined, 1 = convertible. See details.}

\item{ibound}{3D integer array of dimensions \code{dis$nrow x dis$ncol x dis$nlay} corresponding to the MODFLOW grid; specifies the active and inactive cells in the MODFLOW simulation. See details.}

\item{porosity}{numeric array of dimensions \code{dis$nrow x dis$ncol x dis$nlay} specifying the effective porosity for the MODPATH simulation}

\item{bas}{optional \code{RMODFLOW} bas object. See details.}

\item{flow}{optional \code{RMODFLOW} flow object of class \code{lpf}, \code{upw}, \code{bcf} or \code{huf}. See details.}

\item{modflow}{optional \code{RMODFLOW} modflow object. See details.}
}
\value{
Object of class mpbas
}
\description{
\code{rmp_create_bas} creates a \code{RMODPATH} mpbas object.
}
\details{
When a \code{RMODFLOW} bas object is supplied, hdry and ibound are obtained from that object.
 When a \code{RMODFLOW} flow package object (lpf, upw, bcf or huf) is supplied, laytyp and hnoflo are obtained from that object.
 When a \code{RMODFLOW} modflow object is supplied, the \code{RMODFLOW} bas and flow objects are obtained from that object.
}
\seealso{
\code{\link{rmp_read_bas}}, \code{\link{rmp_write_bas}}
}
