% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam.R
\name{rmp_create_nam}
\alias{rmp_create_nam}
\title{Create a RMODPATH mpnam object}
\usage{
rmp_create_nam(
  dis_file,
  mpbas_file,
  head_file,
  budget_file,
  unit_numbers = NULL
)
}
\arguments{
\item{dis_file}{path to the dis file to be used in the MODPATH simulation. Only a MODFLOW-2005 or NWT DIS file is currently supported.}

\item{mpbas_file}{path to the mpbas file to be used in the MODPATH simulation}

\item{head_file}{path to the MODFLOW head output file to be used in the MODPATH simulation}

\item{budget_file}{path to the MODFLOW cell-by-cell budget file to be used in the MODPATH simulation}

\item{unit_numbers}{optional; integer vector of length 4 specifying the unit numbers. Only required for backward compatability with older MODPATH versions.}
}
\value{
Object of class mpnam
}
\description{
\code{rmp_create_nam} creates a \code{RMODPATH} mpnam object.
}
\seealso{
\code{\link{rmp_read_nam}}, \code{\link{rmp_write_nam}}
}
