% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rmp_create_sim}
\alias{rmp_create_sim}
\title{Create a RMODPATH sim object}
\usage{
rmp_create_sim(
  namefilename,
  dis,
  listingfilename = "output.mplist",
  simulationtype = 4,
  trackingdirection = 1,
  weaksinkoption = 1,
  weaksourceoption = 1,
  budgetoutputoption = 1,
  tracemode = 0,
  endpointfilename = "output.endp",
  pathlinefilename = "output.path",
  timeseriesfilename = "output.tms",
  tracefilename = "output.trace",
  traceparticlegroup,
  traceparticelid,
  budgetcellnumbers = NULL,
  referencetimeoption = 1,
  referencetime = 0,
  stressperiod,
  timestep,
  timestepfraction,
  stoptimeoption = 2,
  stoptime,
  timepointoption,
  timepointcount,
  timepointinterval,
  timepoints,
  stopzone,
  zones = NULL,
  retardation = NULL,
  particlegroups
)
}
\arguments{
\item{namefilename}{character; filename of the MODPATH Name File}

\item{dis}{\code{RMODFLOW} dis object; only required when specifying zones or retardation.}

\item{listingfilename}{character; filename of the MODPATH listing file. Defaults to 'output.mplist'.}

\item{simulationtype}{integer specifying the type of MODPATH simulation. Defaults to 4.}

\item{trackingdirection}{integer indicating whether the forward (1; default) or backward (2) particle tracking is performed.}

\item{weaksinkoption}{integer indicating whether particles pass through (1; default) or stop at (2) weak sink cells.}

\item{weaksourceoption}{integer indicating whether particles pass through (1; default) or stop at (2) weak source cells.}

\item{budgetoutputoption}{integer controlling the type of information on individual cell water balance errors to be computed and printed. Defaults to 1 (summary is printed; record headers are not).}

\item{tracemode}{integer specifying whether trace mode is on (1) or off (0; default).}

\item{endpointfilename}{character; name of the MODPATH endpoint output file. Defaults to 'output.endp'.}

\item{pathlinefilename}{character; name of the MODPATH pathline output file. Defaults to 'output.path'.}

\item{timeseriesfilename}{character; name of the MODPATH timeseries output file. Defaults to 'output.tms'.}

\item{tracefilename}{character; name of the MODPATH trace output file. Defaults to 'output.trace'.}

\item{traceparticlegroup}{character; particle group name of the specified particle to be traced when tracemode = 1.}

\item{traceparticelid}{integer; particle ID of the specified particle to be traced when tracemode = 1.}

\item{budgetcellnumbers}{optional integer vector; MODFLOW grid cell numbers for which detailed water budgets are computed.}

\item{referencetimeoption}{integer specifying how the reference time is defined. Default to 1.}

\item{referencetime}{numeric specifying the reference time when referencetimeoption = 1. Defaults to 0.0.}

\item{stressperiod}{integer specifying the  stress period from which the reference time is obtained when referencetimeoption = 2.}

\item{timestep}{integer specifying the time step within stressperiod from which the reference time is obtained when referencetimeoption = 2.}

\item{timestepfraction}{numeric specifying the fraction between 0 and 1 of timestep from which the reference time is obtained when referencetimeoption = 2.}

\item{stoptimeoption}{integer; determines how the simulation is terminated based on time. Defaults to extended the final time step until all particles are terminated (2).}

\item{stoptime}{numeric; tracking time value at which to stop the simulation when stoptimeoption = 3.}

\item{timepointoption}{integer; controls how time point data should be specified when timeseries output is written (simulationtype = 3 or 4).}

\item{timepointcount}{integer; is the number of time points for which time point data should be specified at intervals timepointinterval when timepointoption = 1. When timepointoption = 2, this is determined from the length of timepoints.}

\item{timepointinterval}{numeric; uniform time interval at which time point data is written when timepointoption = 1.}

\item{timepoints}{numeric vector specifying the time points at which to write time point data when timepointoption = 2.}

\item{stopzone}{integer; specifies the zone value in zones where particles are terminated. Only used when zones is also specified. Only positive values are allowed.}

\item{zones}{optional 3d positive integer array of dimensions {dis$nrow x dis$ncol x dis$nlay} specifying the zone numbers for each grid.}

\item{retardation}{optional 3d numeric array of dimensions {dis$nrow x dis$ncol x dis$nlay} specifying the retardation factor for each grid cell.}

\item{particlegroups}{a single \code{RMODPATH} particle_group object or a list of \code{RMODPATH} particle_group objects specifying the particle groups and corresponding starting locations.}
}
\value{
Object of class sim
}
\description{
\code{rmp_create_sim} creates a \code{RMODPATH} sim object.
}
\seealso{
\code{\link{rmp_write_sim}}, \code{\link{rmp_read_sim}}
}
