% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_locations.R
\name{rmp_create_sloc}
\alias{rmp_create_sloc}
\title{Create a RMODPATH starting locations object}
\usage{
rmp_create_sloc(df, inputstyle = 1, ext_file = NULL, ...)
}
\arguments{
\item{df}{data.frame with columns specifying starting location information. See details.}

\item{inputstyle}{integer, type of starting locations input style. Currently, only inputstyle = 1 is supported.}

\item{ext_file}{if \code{NULL} (default), when writing starting locations, the sloc object is written inside the Simulation File. If a character, specifies the path to the external file to which the sloc object will be written.}

\item{...}{ignored}
}
\value{
Object of class sloc which can be used in \code{\link{rmp_create_particlegroup}} to define a particle group's starting locations.
}
\description{
\code{rmp_create_sloc} creates a \code{RMODPATH} sloc object.
}
\details{
For inputstyle = 1, \code{df} should have following columns: (layer, row, column) OR (cellnumber), localx, localy, localz, timeoffset, drape. An optional id column may also be specified.
         \code{\link{rmp_get_locations}} may be used to obtain this data from given x, y and z coordinates.
}
\seealso{
\code{\link{rmp_write_sloc}}, \code{\link{rmp_read_sloc}}, \code{\link{rmp_get_locations}}, \code{\link{rmp_create_particlegroup}}, \code{\link{rmp_create_sim}}
}
