% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_locations.R
\name{rmp_get_locations}
\alias{rmp_get_locations}
\title{Obtain particle starting locations from x, y and z coordinates}
\usage{
rmp_get_locations(
  obj,
  dis,
  prj = RMODFLOW::rmf_get_prj(dis),
  locationstyle = 1,
  timeoffset = NULL,
  drape = NULL,
  id = NULL,
  ...
)
}
\arguments{
\item{obj}{data.frame with x, y and z columns or an {sf}, {sfc} or {sfg} XY or XYZ POINT or MULTIPOINT object. See details.}

\item{dis}{\code{RMODFLOW} dis object}

\item{prj}{optional \code{RMODFLOW} prj object; required when obj contains real-world coordinates instead of MODFLOW grid coordinates}

\item{locationstyle}{integer specifying the locationstyle to be used. The default (1) uses layer, row and column indices; the other (2) a cellnumber index.}

\item{timeoffset}{optional numeric column specifying the timeoffset. This is required by \code{\link{rmp_create_sloc}} when using the outputted tibble as the \code{sloc} argument.}

\item{drape}{optional integer column specifying the drape option. This is required by \code{\link{rmp_create_sloc}} when using the outputted tibble as the \code{sloc} argument.}

\item{id}{optional integer column specifying the particle id. This is an optional argument by \code{\link{rmp_create_sloc}} when using the outputted tibble as the \code{sloc} argument}

\item{...}{ignored}
}
\value{
a tibble with layer, row, column, localx, localy and localz columns when inputstyle = 1 or cellnumber, localx, localy and localz columns when inputstyle = 2,
        corresponding to the given coordinates in obj. When timeoffset and drape (and optionally, id) columns are added, this can be used as the \code{sloc} argument in \code{\link{rmp_create_sloc}} to create a starting locations object.
}
\description{
\code{rmp_get_locations} obtains the starting locations of particles from an object with x, y and z coordinates
}
\details{
When obj is an \code{sf}, \code{sfc} or \code{sfg} object with dimensions XY, an additional z column should be specified. The \code{sf} package should be installed when obj is of class \code{sf}, \code{sfc} or \code{sfg}.
}
\seealso{
\code{\link{rmp_create_sloc}}, \code{\link{rmp_create_particlegroup}}, \code{\link{rmp_create_sim}}
}
