% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bas.R
\name{rmp_read_bas}
\alias{rmp_read_bas}
\title{Read a MODPATH Basic Data File}
\usage{
rmp_read_bas(file, dis, style = "2005", ...)
}
\arguments{
\item{file}{path to the Basic Data file}

\item{dis}{\code{RMODFLOW} dis object}

\item{style}{either '2005', 'usg' or '6' denoting the type of basic file to be read. Currently, only '2005' is supported.}

\item{...}{ignored}
}
\value{
a \code{RMODPATH} mpbas object
}
\description{
\code{rmp_read_bas} reads a MODPATH Basic Data file and returns at as a \code{mpbas} object
}
\seealso{
\code{\link{rmp_write_bas}}, \code{\link{rmp_create_bas}}
}
