% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{rmp_read_endpoint}
\alias{rmp_read_endpoint}
\title{Reads an endpoint output file}
\usage{
rmp_read_endpoint(file)
}
\arguments{
\item{file}{path tot the endpoint output file}
}
\value{
object of class endpoint
}
\description{
\code{rmp_read_endpoint} reads an endpoint file and returns it as a \code{\link{RMODPATH}} endpoint object
}
