% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam.R
\name{rmp_read_nam}
\alias{rmp_read_nam}
\title{Read a MODPATH Name File}
\usage{
rmp_read_nam(file, ...)
}
\arguments{
\item{file}{path to the Name File}

\item{...}{ignored}
}
\value{
a \code{RMODPATH} mpnam object
}
\description{
\code{rmp_read_nam} reads a MODPATH Name File and returns at as a \code{mpnam} object
}
\seealso{
\code{\link{rmp_write_nam}}, \code{\link{rmp_create_nam}}
}
