% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{rmp_read_pathline}
\alias{rmp_read_pathline}
\title{Reads a pathline output file}
\usage{
rmp_read_pathline(file, ...)
}
\arguments{
\item{file}{path tot the pathline output file}

\item{...}{ignored}
}
\value{
object of class pathline
}
\description{
\code{rmp_read_pathline} reads a pathline output file and returns it as a \code{\link{RMODPATH}} pathline object
}
\seealso{
\code{\link{rmp_read_timeseries}}, \code{\link{rmp_read_endpoint}}
}
