% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_locations.R
\name{rmp_read_sloc}
\alias{rmp_read_sloc}
\title{Read starting locations of a particle group}
\usage{
rmp_read_sloc(file = NULL, remaining_lines = NULL, pathname = file, ...)
}
\arguments{
\item{file}{path to the starting locations file. Only necessary when locations are specified in an EXTERNAL file}

\item{remaining_lines}{character vector with remaining lines of the Simulation File. Only necessary when locations are specified using the INTERNAL keyword.}

\item{pathname}{when reading from file, the path specified in the Simulation File, e.g. the basename. Defaults to using \code{file}}

\item{...}{ignored}
}
\value{
If read from \code{file}, a \code{sloc} object containing the starting locations of the particle group. If read from \code{remaining_lines}, a list with the \code{sloc} object and the \code{remaining_lines} of the input vector.
}
\description{
\code{rmp_read_starting_locations} reads the starting locations of a particle group and returns it as a \code{sloc} object
}
\details{
Either \code{file} or \code{remaining_lines} should be specified, but not both.
}
\seealso{
\code{\link{rmp_create_sloc}}, \code{\link{rmp_write_sloc}}
}
