% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{rmp_read_timeseries}
\alias{rmp_read_timeseries}
\title{Reads a timeseries output file}
\usage{
rmp_read_timeseries(file, ...)
}
\arguments{
\item{file}{path tot the timeseries output file}

\item{...}{ignored}
}
\value{
object of class timeseries
}
\description{
\code{rmp_read_timeseries} reads a timeseries output file and returns it as a \code{RMODPATH} timeseries object
}
\seealso{
\code{\link{rmp_read_pathline}}, \code{\link{rmp_read_endpoint}}
}
