% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bas.R
\name{rmp_write_bas}
\alias{rmp_write_bas}
\title{Write a MODPATH Basic Data File}
\usage{
rmp_write_bas(bas, file, ...)
}
\arguments{
\item{bas}{a \code{\link{RMODPATH}} mpbas object}

\item{file}{filename to write to; typically '*.mpbas'}

\item{...}{additional arguments passed to \link{\code{RMODPATH}}{rmpi_write_array}. Can be ignored when arrays are INTERNAL or CONSTANT.}
}
\value{
\code{NULL}
}
\description{
\code{rmp_write_bas} writes a MODPATH Basic Data File based on an \code{\link{RMODPATH}} mpbas object.
}
\seealso{
\code{\link{rmp_read_bas}}, \code{\link{rmp_create_bas}}
}
