% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam.R
\name{rmp_write_nam}
\alias{rmp_write_nam}
\title{Write a MODPATH Name File}
\usage{
rmp_write_nam(nam, file, ...)
}
\arguments{
\item{nam}{a \code{\link{RMODPATH}} mpnam object}

\item{file}{filename to write to; typically '*.mpnam'}

\item{...}{ignored}
}
\value{
\code{NULL}
}
\description{
\code{rmp_write_nam} writes a MODPATH Name File based on an \code{\link{RMODPATH}} mpnam object.
}
\seealso{
\code{\link{rmp_read_nam}}, \code{\link{rmp_create_nam}}
}
