% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rmp_write_sim}
\alias{rmp_write_sim}
\title{Write a MODPATH Simulation File}
\usage{
rmp_write_sim(sim, file, ...)
}
\arguments{
\item{sim}{a \code{RMODPATH} sim object}

\item{file}{filename to write to; typically '*.mpsim'}

\item{...}{additional arguments passed to \code{rmpi_write_array}. Can be ignored when arrays are INTERNAL or CONSTANT.}
}
\value{
\code{NULL}
}
\description{
\code{rmp_write_sim} writes a MODPATH Simulation File based on an \code{RMODPATH} sim object.
}
\seealso{
\code{\link{rmp_read_sim}}, \code{\link{rmp_create_sim}}
}
