% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_locations.R
\name{rmp_write_sloc}
\alias{rmp_write_sloc}
\title{Write starting locations of a particle group}
\usage{
rmp_write_sloc(
  sloc,
  file,
  append = sloc$startinglocationsfileoption == "INTERNAL",
  ...
)
}
\arguments{
\item{sloc}{a \code{\link{RMODPATH}} sloc object}

\item{file}{path to write to}

\item{append}{logical, should the file be appended? Defaults to TRUE}

\item{...}{ignored}
}
\value{
\code{NULL}
}
\description{
\code{rmp_write_starting_locations} writes the starting locations of a particle group based on an \code{\link{RMODPATH}} sloc object to either a Simulation File or an external ASCII file.
}
\seealso{
\code{\link{rmp_create_sloc}}, \code{\link{rmp_read_sloc}}
}
