% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmpi_parse_comments}
\alias{rmpi_parse_comments}
\title{Read comments
Internal function used in the read_* functions to read comments}
\usage{
rmpi_parse_comments(remaining_lines)
}
\description{
Read comments
Internal function used in the read_* functions to read comments
}
\details{
removes empty comments and prevents copying of RMODPATH header comment
}
\keyword{internal}
