% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmpi_parse_variables}
\alias{rmpi_parse_variables}
\title{Read modpath variables
If all are numbers, returns numeric, otherwise returns character vector}
\usage{
rmpi_parse_variables(remaining_lines, format = "free", ...)
}
\arguments{
\item{format}{character, either \code{'free'} (default) or \code{'fixed'}. When 'fixed', reads 10-character fields and converts to numeric. Empty fields are set to zero.}

\item{...}{additional arguments passed to \link[RMODFLOW]{rmfi_parse_variables}}
}
\description{
Read modpath variables
If all are numbers, returns numeric, otherwise returns character vector
}
\keyword{internal}
