% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmpi_write_array}
\alias{rmpi_write_array}
\title{Write MODPATH array
Internal function used in the rmp_write_* functions for writing array datasets}
\usage{
rmpi_write_array(array, file, ...)
}
\arguments{
\item{...}{additional arguments passed to \link[RMODFLOW]{rmfi_write_array}}
}
\description{
Write MODPATH array
Internal function used in the rmp_write_* functions for writing array datasets
}
\details{
if the array should be written as integers, an integer array should be provided
}
\keyword{internal}
