% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmpi_write_variables}
\alias{rmpi_write_variables}
\title{Write MODPATH variables
Internal function used in the rmp_write_* functions for writing variables}
\usage{
rmpi_write_variables(
  ...,
  file,
  append = TRUE,
  width = 10,
  format = "free",
  integer = FALSE,
  iprn = -1
)
}
\description{
Write MODPATH variables
Internal function used in the rmp_write_* functions for writing variables
}
\keyword{internal}
