% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_locations.R
\name{rmp_create_particlegroup}
\alias{rmp_create_particlegroup}
\title{Create a RMODPATH particle group object}
\usage{
rmp_create_particlegroup(
  sloc,
  name,
  releaseoption,
  releasetimes,
  timecount,
  initialreleasetime,
  interval
)
}
\arguments{
\item{sloc}{\code{RMODPATH} sloc object specyfing the starting locations as obtained by \code{\link{rmp_create_sloc}}.}

\item{name}{character, particle group name}

\item{releaseoption}{integer specifying the release option}

\item{releasetimes}{single numeric value when releaseoption = 1 or numeric vector when releaseoption = 3; specifies the releasetimes.}

\item{timecount}{numeric; number of multiple releasetimes when releaseoption = 2. When releaseoption = 3, this is obtained from the length of releasetimes.}

\item{initialreleasetime}{numeric; value of tracking time at which first particles are released when releaseoption = 2.}

\item{interval}{numeric; uniform time interval between particle releases when releaseoption = 2.}
}
\value{
an object of class particle_group to be used by \code{\link{rmp_create_sim}}
}
\description{
\code{rmp_create_particlegroup} creates a \code{RMODPATH} particle_group object.
}
\seealso{
\code{\link{rmp_create_sim}}, \code{\link{rmp_create_sloc}}, \code{\link{rmp_get_locations}}
}
